refactor(be): decouple verified-emails from email_recovery
The inbound-DKIM challenge primitive (pending-challenge map, SMTP
gateway dispatcher, DKIM leaf submission, nonce RNG, shared
prepare-common core, shared constants) was nested inside the recovery
module even though both the recovery flow and the new verified-email
flow consume it equally. Hoist it into a sibling top-level module so
the two flows are truly parallel and neither imports from the other.

* New module `src/internet_identity/src/email_inbound/` owns the
  shared primitive: pending.rs, rng.rs, smtp.rs, submit_leaf.rs (whole
  files, moved via `git mv`), plus a new prepare.rs holding the
  extracted `prepare_common` and DNSSEC/normalize helpers, and a new
  mod.rs collecting the shared constants and wrapper helpers.
* `email_recovery/` slimmed to the recovery-flow surface only:
  `prepare_add`, `prepare_delegation`, `remove_credential`,
  `recovery_seed_for_nonce`. A `pub use crate::email_inbound::{...}`
  facade keeps external callers (dkim/, dmarc/, doc references,
  main.rs) source-compatible through this PR; follow-up will migrate
  them to import directly.
* `verified_emails/` (previously nested inside email_recovery/) is now
  a crate-root sibling of email_recovery/, importing only from
  email_inbound/, with its own RemoveError so the two flow modules
  share no types either.
* External callers `session_delegation::email_seed` and
  `authz_utils::email_recovery_seed` migrated to
  `crate::email_inbound::smtp::calculate_email_recovery_seed`
  (the pub(crate) submodule can't be re-exported pub through the
  email_recovery facade).
