+++ src/internet_identity/src/email_inbound/submit_leaf.rs
@@ -48,9 +48,8 @@
 //!
 //! See design doc §8.4 / §8.5.
 
-use super::pending::{PendingKind, PendingStatus};
+use super::pending::{self, PendingKind, PendingStatus};
 use crate::doh::{DohError, DohRecord};
-use crate::email_recovery::pending;
 use crate::single_flight_cache::Cached;
 use internet_identity_interface::internet_identity::types::email_recovery::{
     EmailRecoveryError, EmailRecoverySubmitDkimLeafArg,
