+++ src/internet_identity/src/session_delegation.rs
@@ -188,7 +188,7 @@ mod tests {
 
         let session_seed = session_delegation_seed(anchor_number);
 
-        let email_seed = crate::email_recovery::smtp::calculate_email_recovery_seed(
+        let email_seed = crate::email_inbound::smtp::calculate_email_recovery_seed(
             "user@example.com",
             anchor_number,
         );
